Folders Organization ; Player sprites update + idle sprites added Update all scripts links
