+++ PhantomProjects/Boss_/Boss.cs
@@ -5,8 +5,9 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 using PhantomProjects.States;
-using PhantomProjects.PlayerBullets;
+using PhantomProjects.Player_;
 using Microsoft.Xna.Framework.Content;
+using PhantomProjects.Map_;
 
 namespace PhantomProjects.Boss_
 {
