+++ PhantomProjects/Boss_/FireballManager.cs
@@ -5,7 +5,10 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
-using PhantomProjects.PlayerBullets;
+using PhantomProjects.Player_;
+using PhantomProjects.Explosion_;
+using PhantomProjects.Map_;
+
 
 namespace PhantomProjects.Boss_
 {
