+++ PhantomProjects/Enemy_/BulletE.cs
@@ -5,7 +5,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 
-namespace PhantomProjects
+namespace PhantomProjects.Enemy_
 {
     class BulletE
     {
