+++ PhantomProjects/Enemy_/EnemyManager.cs
@@ -5,9 +5,10 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
-using PhantomProjects.PlayerBullets;
+using PhantomProjects.Player_;
+using PhantomProjects.Explosion_;
 
-namespace PhantomProjects
+namespace PhantomProjects.Enemy_
 {
     class EnemyManager
     {
