+++ PhantomProjects/Interactables_/Keycard.cs
@@ -2,12 +2,12 @@
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
-using PhantomProjects.PlayerBullets;
+using PhantomProjects.Player_;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace PhantomProjects.Interactables
+namespace PhantomProjects.Interactables_
 {
     class Keycard
     {
