+++ PhantomProjects/Map_/RectangleHelper.cs
@@ -5,7 +5,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 
-namespace PhantomProjects
+namespace PhantomProjects.Map_
 {
     static class RectangleHelper
     {
